fix: bump psa version (#446)
- bump the PSA to latest verison (i.e. v6.0.0)
- Update the `upload-artifact` and `download-artifact` action version to
v4 since v3 is deprecated now.
- Correction in the HEC token value being passed when executing the
tests. HEC token value was different in pytest.ini and in the
run_splunk.sh file resulting in failure of HEC validation when executing
the tests. It didn't failed the tests earlier because the PSA v5.2.6 and
earlier didn't had the HEC token validation.

test run:
https://github.com/splunk/splunk-add-on-for-amazon-web-services/actions/runs/13941331735